refactor(diagram): lift the point math out of the SVG front end
Curve sampling, shape classification and the arc-length midpoint are not
SVG-specific: they operate on a run of points and answer questions any
vector source has to ask. Only the walk over the source differs.

Moving them into a `polyline` module leaves `svg/geometry` with just
the usvg walk, and gives a second front end somewhere to arrive.

No behaviour change.
